reorganised resources dir for stapler
Fixed the categories showing up, when you click save on Jenkins now the page crashes. Something to do with saving the category item.
