+++ src/main/java/io/jenkins/plugins/OnboardingTask/Category.java
@@ -1,13 +1,13 @@
-package io.jenkins.plugins;
+package io.jenkins.plugins.OnboardingTask;
 
 import java.util.UUID;
 
-import org.kohsuke.stapler.DataBoundConstructor;
-
 import hudson.Extension;
 import hudson.ExtensionList;
+import org.kohsuke.stapler.DataBoundConstructor;
 import hudson.model.AbstractDescribableImpl;
 import hudson.model.Descriptor;
+
 
 public class Category extends AbstractDescribableImpl<Category> {
 
@@ -16,11 +16,6 @@ public class Category extends AbstractDescribableImpl<Category> {
 
     @DataBoundConstructor
     public Category(String categoryName, UUID uuid) {
-        this.categoryName = categoryName;
-        this.uuid = uuid;
-    }
-
-    public Category(String categoryName) {
         this.categoryName = categoryName;
         this.uuid = java.util.UUID.randomUUID();
     }
