--- src/main/resources/io.jenkins.plugins.Category/config.jelly
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
-
-    <f:entry field="categoryName" title="${%categoryName}">
-        <f:textbox/>
-    </f:entry>
-
-    <f:entry field="uuid" title="${%uuid}">
-        <f:textbox/>
-    </f:entry>
-
-</j:jelly>
