+++ src/main/resources/io/jenkins/plugins/OnboardingTask/Category/config.jelly
@@ -1,0 +1,13 @@
+
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+
+    <f:entry field="${it.categoryName}" title="${%category name}">
+        <f:textbox/>
+    </f:entry>
+
+    <f:entry field="${it.uuid}" title="${%uuid}">
+        <f:readOnlyTextbox/>
+    </f:entry>
+
+</j:jelly>
