+++ src/main/resources/io/jenkins/plugins/OnboardingTask/index.jelly
