+++ src/test/java/io/jenkins/plugins/OnboardingTask/OnboardingTaskTest.java
@@ -1,4 +1,4 @@
-package io.jenkins.plugins;
+package io.jenkins.plugins.OnboardingTask;
 
 import com.gargoylesoftware.htmlunit.html.HtmlForm;
 import com.gargoylesoftware.htmlunit.html.HtmlTextInput;
